requirements.txt
alembic==1.12.1
argon2-cffi==25.1.0
boto3==1.34.162
bcrypt==4.1.2
email-validator==2.3.0
fastapi==0.121.3
httpx==0.28.1
mysql-connector-python==9.4.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.5.0
pyjwt==2.12.0
pymysql==1.1.1
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.26
sqlalchemy==2.0.23
twilio==8.10.0
uvicorn==0.24.0